Backend/AIArch/requirements.txt
fastapi[all]
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
